# backend/app/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Stay Sidekick API
  version: 1.0.0
  description: |
    API REST del sistema de gestión hotelera Stay Sidekick.

    ## Autenticación
    La mayoría de endpoints requieren un JWT en la cabecera `Authorization: Bearer <token>`.
    El token se obtiene en `POST /api/auth/login`.

    Los endpoints de escritura también exigen el **CSRF Double-Submit Cookie**:
    obten el token en `GET /api/csrf-token` y envíalo en la cabecera `X-CSRF-Token`.

    ## Roles
    - **operativo** — acceso a herramientas, solo lectura de configuración.
    - **admin** — gestión completa de la empresa (usuarios, integraciones, configuración).
    - **superadmin** — acceso transversal a todas las empresas.
servers:
  - url: http://localhost:5000
    description: Desarrollo local (Docker)
  - url: https://nginx-sidekick.railway.app
    description: Producción (Railway)
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CsrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    Ok:
      type: object
      properties:
        ok:
          type: boolean
          example: true
    Error:
      type: object
      properties:
        ok:
          type: boolean
          example: false
        errors:
          type: array
          items:
            type: string
...
    PlantillaVault:
      type: object
      properties:
        id:
          type: string
          format: uuid
        nombre:
          type: string
        contenido:
          type: string
        idioma:
          type: string
          enum: [es, en, fr, de, it, pt]
        categoria:
          type: string
          nullable: true
          enum: [BIENVENIDA, INSTRUCCIONES, RECORDATORIO, CHECKIN_TARDIO, CHECKOUT, INCIDENCIA, GENERAL]
        activa:
          type: boolean
...
paths:
...
  /api/vault/plantillas:
    get:
      tags: [Vault]
      summary: Listar plantillas de comunicación
      security:
        - BearerAuth: []
      parameters:
        - name: categoria
          in: query
          schema:
            type: string
            enum: [BIENVENIDA, INSTRUCCIONES, RECORDATORIO, CHECKIN_TARDIO, CHECKOUT, INCIDENCIA, GENERAL]
        - name: idioma
          in: query
          schema:
            type: string
            enum: [es, en, fr, de, it, pt]
      responses:
        '200':
          description: Lista de plantillas
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  plantillas:
                    type: array
                    items:
                      $ref: '#/components/schemas/PlantillaVault'

    post:
      tags: [Vault]
      summary: Crear plantilla
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, contenido]
              properties:
                nombre:
                  type: string
                  maxLength: 200
                contenido:
                  type: string
                idioma:
                  type: string
                  enum: [es, en, fr, de, it, pt]
                  default: es
                categoria:
                  type: string
                  enum: [BIENVENIDA, INSTRUCCIONES, RECORDATORIO, CHECKIN_TARDIO, CHECKOUT, INCIDENCIA, GENERAL]
                  nullable: true
      responses:
        '201':
          description: Plantilla creada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  plantilla:
                    $ref: '#/components/schemas/PlantillaVault'
  /api/vault/plantillas/{plantilla_id}:
    put:
      tags: [Vault]
      summary: Actualizar plantilla
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlantillaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, contenido]
              properties:
                nombre:
                  type: string
                contenido:
                  type: string
                idioma:
                  type: string
                categoria:
                  type: string
                  nullable: true
      responses:
        '200':
          description: Plantilla actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  plantilla:
                    $ref: '#/components/schemas/PlantillaVault'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Vault]
      summary: Eliminar plantilla (soft-delete)
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlantillaId'
      responses:
        '200':
          description: Plantilla eliminada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ok'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/vault/plantillas/{plantilla_id}/mejoras:
    post:
      tags: [Vault]
      summary: Mejorar plantilla con IA
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlantillaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contenido, idioma]
              properties:
                contenido:
                  type: string
                idioma:
                  type: string
                  enum: [es, en, fr, de, it, pt]
      responses:
        '200':
          description: Texto mejorado por IA
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  contenido:
                    type: string
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          description: Límite de IA alcanzado
  /api/vault/plantillas/{plantilla_id}/traducciones:
    post:
      tags: [Vault]
      summary: Traducir plantilla con IA
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PlantillaId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [contenido, idioma_destino]
              properties:
                contenido:
                  type: string
                idioma_destino:
                  type: string
                  enum: [es, en, fr, de, it, pt]
      responses:
        '200':
          description: Texto traducido
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  contenido:
                    type: string
        '429':
          description: Límite de IA alcanzado
  /api/ai/uso:
    get:
      tags: [Vault]
      summary: Contadores de uso de IA
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Uso diario y semanal
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  uso_hoy:
                    type: integer
                  uso_semana:
                    type: integer
                  limite_diario:
                    type: integer
                  limite_semanal:
                    type: integer
  /api/ai/config:
    get:
      tags: [Vault]
      summary: Configuración de IA enmascarada (admin)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Proveedor, modelo y api_key enmascarada
          content:
            application/json:
              schema:
                type: object

  # ── Notificaciones ────────────────────────────────────────────────────────
...
components:
  parameters:
...
    PlantillaId:
      name: plantilla_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...